sve: add registers to arch lists
This does _not_ properly handle or deal with SVE registers, it merely
makes things less broken when SVE registers are present. Currently the
disassembler defines the register list, which the lifter users to emit
LLIL. Because the architecture was not aware of SVE registers, any
instructions that references them that were lifted would emit references
to registers that didn't exist as far as the architecture was concerned.
This would break in weird ways, see
https://github.com/Vector35/binaryninja-api/issues/2307

This PR adds the registers to the register list with the following
behavior:
- All variable length SVE registers (Z*) are 128bits wide and alias the V
  registers
- All predicate registers (P*) are 256 bits wide, and are their own regsiter
- All PF registers are 0 bits wide. I can find no documentation on what
  these are.

Fixes #80