use V registers instead of Q
As part of the vector register refactor we made the V registers the
"base" register, which are windowed into by the b, h, s, and q
registers. However, the calling convention was not updated to switch
from q to v registers, nor were the full width registers changed. The
v and q registers are 1:1 aliases of each other, both 16 bytes.

Currently, the situation is kind of weird: the Q registers are defined
as the full width registers, however are implemented in terms of a view
into the V registers. This is probably a bug, even though it mostly
works -- I think (?) the API expects that you'll always window into the
full width regsters. The fix for this is easy as theyre both the same
width -- we just make the V regsiters the full width regsiters.

Next, we just update the calling convetion to use the V registers
instead of their Q counterparts. This is partially due to calling
conventions inability to use subregisters (even though V and Q are the
same size, Q still isnt a full width register...), see
https://github.com/Vector35/binaryninja-api/issues/2793. However, even
aside from this issue, this appears to be more accurate:
- https://developer.arm.com/documentation/den0024/a/The-ABI-for-ARM-64-bit-Architecture/Register-use-in-the-AArch64-Procedure-Call-Standard/Parameters-in-NEON-and-floating-point-registers
- https://docs.microsoft.com/en-us/cpp/build/arm64-windows-abi-conventions?view=msvc-170
- https://en.wikipedia.org/wiki/Calling_convention#ARM_(A64)

With this change applied, functions that take floating point paramters
are cleaned up, e.g. `atan` goes from `void atan(double arg1 @ v0)` to
`void atan(double arg1)`.