fix movk
movk overwrites a 16bit slice of a register with an immediate from the
instruction. The current lift assumes the existing bits are zero and
just ORs in the new bits. This is incorrect and results in dataflow
inaccuracies.

Theoretically we could introduce even more register windows to handle
this instead of masking but ehhhhhhhhh....

Reported by @ek0